app/requirements.txt
streamlit
torch==1.13.1
torchvision==0.14.1
facenet-pytorch
numpy
Pillow
opencv-python
scikit-learn